# Conflicts:
#	pom.xml
#	src/main/java/com/cryptowallet/entities/Address.java
#	src/main/java/com/cryptowallet/entities/Role.java
#	src/main/java/com/cryptowallet/entities/User.java
#	src/main/java/com/cryptowallet/repositories/UserRepository.java
#	src/main/java/com/cryptowallet/services/interfaces/UserService.java
#	src/main/resources/application.properties
#	src/main/resources/templates/index.html
#	src/test/resources/application-test.properties

+++ src/main/java/com/cryptowallet/repositories/UserRepository.java
@@ -1,14 +1,14 @@
 package com.cryptowallet.repositories;
 
 import com.cryptowallet.entities.User;
+import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     boolean existsByLogin(@NonNull String login);
 
     boolean existsByEmail(@NonNull String email);
